social-network-hw3/post-core-service/requirements.txt
grpcio
grpcio-tools
confluent-kafka
